requirements.txt
deep-translator
tqdm


# CLIP baseline requirements
# Install with: pip install -r requirements_clip.txt

torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/openai/CLIP.git
Pillow>=9.0.0
tqdm>=4.65.0